In the property accessor, if the object supports the sequence or mapping protocol, just copy the container's contents into the output hashtable.
This isn't strictly correct; it's a side-effect of our approach of
treating all Python objects as PHP objects.  We'd really like PHP to
attempt to attempt to "cast" our Python object to an array-like object
first and then use the dimension APIs, in the spirit of Python's "duck
typing", but those casting operations don't currently exist.  We now have
the potential for false-positive conversions below, where we return the
sequence or mapping contents of a Python object that also has a legitimate
set of additional properties.


git-svn-id: http://svn.php.net/repository/pecl/python@241628 c90b9560-bf6c-de11-be94-00142212c4b1